# openapi/sandicts-api.json
{
  "openapi": "3.1.0",
  "paths": {
    "/auth/magic-link/request": {
      "post": {
        "description": "Accepts a magic link request without revealing account existence.",
        "operationId": "RequestMagicLinkController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestMagicLinkBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Magic link request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestMagicLinkResponse_Output"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Codes: `validation_error`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        400
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "validation_error"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Codes: `rate_limited`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        429
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "rate_limited"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. Codes: `internal_error`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        500
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "internal_error"
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Required dependency is unavailable. Codes: `email_delivery_unavailable`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        503
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "email_delivery_unavailable"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Request magic link",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "description": "Rotates the refresh token cookie and returns a new access token.",
        "operationId": "RefreshAuthSessionController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Refreshed authenticated session response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshAuthSessionResponse_Output"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed. Codes: `invalid_refresh_token`, `refresh_token_expired`, `refresh_token_reused`, `refresh_token_revoked`, `auth_session_inactive`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        401
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "invalid_refresh_token",
                        "refresh_token_expired",
                        "refresh_token_reused",
                        "refresh_token_revoked",
                        "auth_session_inactive"
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access forbidden. Codes: `account_auth_forbidden`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        403
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "account_auth_forbidden"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Codes: `rate_limited`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        429
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "rate_limited"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. Codes: `internal_error`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        500
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "internal_error"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Refresh auth session",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/sign-out-all": {
      "post": {
        "description": "Revokes all active auth sessions for the authenticated account and clears the refresh cookie for the current browser.",
        "operationId": "SignOutAllController_signOutAllSessions",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer access token for the current account.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "All account sessions signed out"
          },
          "401": {
            "description": "Authentication failed. Codes: `invalid_access_token`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        401
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "invalid_access_token"
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Codes: `rate_limited`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        429
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "rate_limited"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal failure. Codes: `internal_error`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "enum": [
                        500
                      ]
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "internal_error"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Sign out all sessions",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sandicts API",
    "description": "API documentation generated from Nest controllers and Zod schemas.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ApiValidationIssue": {
        "type": "object",
        "properties": {
          "path": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            }
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ],
        "id": "ApiValidationIssue"
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "bad_request",
              "validation_error",
              "rate_limited",
              "unauthorized",
              "invalid_google_credential",
              "invalid_magic_link_token",
              "magic_link_expired",
              "magic_link_already_used",
              "magic_link_superseded",
              "email_delivery_unavailable",
              "invalid_access_token",
              "invalid_refresh_token",
              "refresh_token_expired",
              "refresh_token_reused",
              "refresh_token_revoked",
              "auth_session_inactive",
              "forbidden",
              "account_auth_forbidden",
              "external_identity_conflict",
              "resource_not_found",
              "conflict",
              "business_rule_violation",
              "internal_error"
            ]
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "requestId": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiValidationIssue"
            }
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "requestId"
        ],
        "id": "ApiErrorResponse"
      },
      "RequestMagicLinkBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ],
        "id": "RequestMagicLinkBody"
      },
      "RequestMagicLinkResponse_Output": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "const": "accepted"
          }
        },
        "required": [
          "status"
        ],
        "id": "RequestMagicLinkResponse_Output",
        "additionalProperties": false
      },
...
      "AuthAccountResponse_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ],
        "additionalProperties": false,
        "id": "AuthAccountResponse_Output"
      },
      "AuthSessionResponse_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": false,
        "id": "AuthSessionResponse_Output"
      },
...
      "RefreshAuthSessionResponse_Output": {
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AuthAccountResponse_Output"
          },
          "session": {
            "$ref": "#/components/schemas/AuthSessionResponse_Output"
          },
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "account",
          "session",
          "accessToken",
          "accessTokenExpiresAt"
        ],
        "id": "RefreshAuthSessionResponse_Output",
        "additionalProperties": false
      },
...
    }
  }
}